+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fix nextList returning true for single-item lists #108
 - Fix broken range calculations for the end of a list of ranges #113
+- `\Lullabot\Mpx\Service\IdentityManagement\UserInterface` has been added to
+  simplify bridging user configuration from applications. #116
 
 ### Removed